ditched "maven structure"
Moved code around, fixed builds to be grunt-only, added .travis.yml.
Switched default karma run to use PhantomJS so it can be headless on
travis.
