+++ .gitignore
@@ -1,1 +1,3 @@
 ./r.js
+
+/node_modules
