+++ Gruntfile.js
@@ -4,7 +4,7 @@ module.exports = function (grunt) {
 
     grunt.initConfig({
         jshint: {
-            src: ['src/main/**/*.js', 'src/test/**/*.js', '!node_modules/**/*.*'],
+            src: ['js/**/*.js', 'test/**/*.js', '!node_modules/**/*.*'],
             options: {
                 jshintrc: '.jshintrc'
             }
@@ -20,15 +20,29 @@ module.exports = function (grunt) {
                 files: ['**/*.js', '!**/node_modules/**'],
                 tasks: ['lint', 'test']
             }
+        },
+        requirejs: {
+            compile: {
+                options: {
+                    baseUrl: "./js",
+                    paths: {
+                        bullhorn: "."
+                    },
+                    name: "bullhorn/ChannelFactory",
+                    out: "bullhorn-min.js"
+                }
+            }
         }
     });
 
     grunt.loadNpmTasks('grunt-contrib-jshint');
     grunt.loadNpmTasks('grunt-contrib-watch');
+    grunt.loadNpmTasks('grunt-contrib-requirejs');
     grunt.loadNpmTasks('grunt-karma');
 
-    grunt.registerTask('test', ['karma']);
     grunt.registerTask('lint', ['jshint']);
-    grunt.registerTask('default', ['jshint', 'karma']);
+    grunt.registerTask('test', ['karma']);
+    grunt.registerTask('compile', ['requirejs']);
+    grunt.registerTask('default', ['lint', 'test']);
 
 };
