+++ examples/require.js
