+++ examples/validate.js
