+++ karma.conf.js
@@ -14,9 +14,9 @@ module.exports = function(config) {
 
     // list of files / patterns to load in the browser
     files: [
-        'src/test/test-main.js',
-        {pattern: 'src/main/**/*.js', included: false},
-        {pattern: 'src/test/**/*.js', included: false}
+        'test/test-main.js',
+        {pattern: 'js/**/*.js', included: false},
+        {pattern: 'test/**/*.js', included: false}
     ],
 
 
@@ -54,7 +54,7 @@ module.exports = function(config) {
     // - Safari (only Mac; has to be installed with `npm install karma-safari-launcher`)
     // - PhantomJS
     // - IE (only Windows; has to be installed with `npm install karma-ie-launcher`)
-    browsers: ['Chrome'],
+    browsers: ['PhantomJS'],
 
 
     // If browser does not capture in given timeout [ms], kill it
@@ -65,6 +65,7 @@ module.exports = function(config) {
         'karma-mocha',
         'karma-chai',
         'karma-chrome-launcher',
+        'karma-phantomjs-launcher',
         'karma-mocha-reporter'
     ],
