--- pom.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-This is a simple POM file to generate a jar file containing the library for publishing in a maven repository. It can be used in the following ways.
- - "mvn install" will add the production of a jar maven artifact containing the library and publish it to your local maven repository
- - "mvn deploy" will add publishing to a remote maven repository, if you define the repository and have the correct access to it.
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-	<modelVersion>4.0.0</modelVersion>
-	<groupId>com.atsid</groupId>
-	<artifactId>bullhorn-js</artifactId>
-	<version>1.0.4</version>
-	<packaging>jar</packaging>
-
-	<!-- PROJECT PROPERTIES -->
-	<properties>
-		<javascript.source>${basedir}/src/main</javascript.source>
-	</properties>
-
-    <distributionManagement>
-        <repository>
-            <id>your-repo</id>
-            <name>Your Repo</name>
-            <url>https://your-repo.host.com/repo</url>
-        </repository>
-    </distributionManagement>
-
-	<build>
-		<resources>
-			<resource>
-				<directory>${javascript.source}</directory>
-			</resource>
-		</resources>
-		<plugins>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-jar-plugin</artifactId>
-				<version>2.4</version>
-				<configuration>
-					<classesDirectory>${javascript.source}</classesDirectory>
-				</configuration>
-			</plugin>
-		</plugins>
-	</build>
-</project>
