--- single-file-build.js
@@ -1,11 +1,0 @@
-/**
- * Build profile for RequireJs' optimizer.
- */
-({
-    baseUrl: "./src/main",
-    paths: {
-        bullhorn: "."
-    },
-    name: "bullhorn/ChannelFactory",
-    out: "bullhorn-min.js"
-})
