--- src/examples/BullhornExample.html
@@ -1,94 +1,0 @@
-<!--
-Simple sample page to demonstrate the use of topic-based messaging with bullhorn.
-It retrieves three instances of a channel defined by a schema, publishes on one and
-subscribes on the other two.
-It is intended to be loaded directly from the file system in this directory.
-Most of the code is setting up AMD (which becomes negligible in larger projects).
--->
-<!DOCTYPE html>
-<html>
-<head>
-    <!--
-    Include only an amd loader.
-    -->
-    <script src="./require.js"></script>
-    <script>
-        // Configure AMD to find things.
-        var path = location.pathname.replace(/\/[^/]+$/, '');
-        require.config({
-            baseUrl: "./",
-            paths: {
-                bullhorn: "../main", // schematic-js code
-                schema: "schema", // the example schema
-                example: "."
-            },
-            // remap to use a third party validator.
-            map: {
-                "*": {
-                    "bullhorn/Validator": "example/Validator"
-                }
-            },
-            config: {
-                "bullhorn/log": {
-                    logLevel: "debug"
-                }
-            }
-        });
-    </script>
-    <script>
-        // Use AMD to define and load the resources we depend on.
-        require([
-            "bullhorn/ChannelFactory",
-            "schema/ExampleChannel"
-        ], function (
-                ChannelFactory,
-                ExampleChannel
-        ) {
-            // Create a channel factory.
-            var cfact = new ChannelFactory({
-                    // "resolver" is used by the factory to find schemas based on names.
-                    // resolve schema by mapping it to the dependency where it was already loaded.
-                    resolver: function (name) {
-                        if (name === "schema/ExampleChannel") {
-                            return ExampleChannel;
-                        }
-                    }
-                }),
-                sub1 = document.getElementsByName("FirstSubscriber")[0],
-                sub2 = document.getElementsByName("SecondSubscriber")[0],
-                msg = document.getElementsByName("SenderText")[0],
-                btn = document.getElementsByName("Sender")[0],
-
-                // Retrieve channels for each "scope" wanting to operate on the channel
-                originatingChannel = cfact.get("schema/ExampleChannel", msg),
-                receiver1Channel = cfact.get("schema/ExampleChannel", sub1),
-                receiver2Channel = cfact.get("schema/ExampleChannel", sub2);
-
-            // Channels are "scoped" based on the second argument of the
-            // factory's get method, so subscriber callbacks are called in
-            // that scope.
-            receiver1Channel.subscribe(function (message) {
-                this.innerHTML = message.text;
-            });
-            receiver2Channel.subscribe(function (message) {
-                this.innerHTML = message.text;
-            });
-
-            // publish a message on the channel, receivers should get it.
-            btn.onclick = function (evt) {
-                originatingChannel.publish( { text: msg.value});
-            }
-
-         });
-    </script>
-
-</head>
-<body>
-    <button name="Sender">Send</button>
-    <textarea name="SenderText"></textarea>
-    <div>FirstSubscriber</div>
-    <div name="FirstSubscriber">Empty</div>
-    <div>SecondSubscriber</div>
-    <div name="SecondSubscriber">Empty</div>
-</body>
-</html>
