--- src/examples/schema/ExampleChannel.js
@@ -1,13 +1,0 @@
-define({
-    "id": "schema/ExampleChannel",
-    "description": "A simple channel for testing",
-    "$schema": "http://json-schema.org/draft-03/schema",
-    "type": "object",
-    "properties": {
-        "text": {
-            "type": "string",
-            "description": "a property.",
-            "required": true
-        }
-    }
-});
