--- src/main/package.json
@@ -1,28 +1,0 @@
-{
-    "name": "bullhorn-js",
-    "version": "1.0.4",
-    "description": "A client-side library for global topic-based messaging.",
-    "keywords": ["client-side", "messaging", "Publish/Subscribe", "channels", "JSON Schema", "json-schema", "schema", "Topic-based Messaging", "pubsub"],
-    "homepage": "http://atsid.github.io/",
-    "licenses": [
-        {
-            "type": "Apache2.0",
-            "url": "http://www.apache.org/licenses/LICENSE-2.0"
-        }
-    ],
-    "contributors": [
-        {
-            "name": "natoverse"
-        },
-        {
-            "name": "tom-servo"
-        }
-    ],
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/atsid/bullhorn-js.git"
-    },
-    "dojoBuild": "profile.js",
-    "dependencies": {
-	}
-}
