+++ test/MockChannel.js
